Update lots of dirty-flag table entries

+++ GPU/Directx9/GPU_DX9.h
@@ -76,7 +76,7 @@ public:
 
 	typedef void (GPU_DX9::*CmdFunc)(u32 op, u32 diff);
 	struct CommandInfo {
-		u8 flags;
+		uint64_t flags;
 		GPU_DX9::CmdFunc func;
 	};
